requirements.txt
# Google Cloud Dialogflow CX
google-cloud-dialogflow-cx
python-dotenv
